+++ app/Models/UserSkill.php
@@ -6,5 +6,7 @@ use Illuminate\Database\Eloquent\Model;
 
 class UserSkill extends Model
 {
-    //
+    protected $hidden = [
+        'created_at', 'updated_at', 'user_id', 'id', 'skill_steps_id'
+    ];
 }