+++ database/migrations/2026_03_15_084431_create_user_skills_table.php
@@ -13,6 +13,10 @@ return new class extends Migration
     {
         Schema::create('user_skills', function (Blueprint $table) {
             $table->id();
+            $table->foreignId('user_id')->constrained()->cascadeOnDelete();
+            $table->foreignId('skill_steps_id')->constrained()->cascadeOnDelete();
+            $table->boolean('is_completed')->default(false);
+            $table->timestamp('completed_at')->nullable();
             $table->timestamps();
         });
     }